% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actogram.R
\name{actogram}
\alias{actogram}
\title{Plot a an single actogram over a period of time for a specific variable}
\usage{
actogram(
  df,
  activity,
  start,
  end,
  save = FALSE,
  outputdir = "testresults",
  outplotname = "actoplot",
  width = 10,
  device = "tiff",
  height = 5
)
}
\arguments{
\item{df}{The dataframe containing the activity data}

\item{activity}{the name of activity}

\item{start}{The start day (in "\%Y\%m\%d" format)}

\item{end}{The end day (in "\%Y\%m\%d" format)}

\item{save}{TRUE if you want to save the image (if TRUE, subsequent arguments
should be provided)}

\item{outputdir}{the directory where you wish to save the image}

\item{outplotname}{the name of the output plot (only name, without device
extension nor path)}

\item{width}{width of the image in cm}

\item{device}{one of 'tiff', 'png', 'pdf', 'jpg'}

\item{height}{height of the image in cm}
}
\value{
None
}
\description{
Takes an activity dataset as input and plot and save an actogram of the
specified activity column
}
