% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df603}
\alias{df603}
\title{df603 Activity Data Sets}
\format{
A data frame of 3 columns
\describe{
\item{datetime}{a POSIX formatted datetime}
\item{move_x}{The acceleration along the x axis}
\item{move_y}{The acceleration along the y axis}
}
}
\source{
Agroscope Posieux
}
\usage{
df603
}
\description{
A dataset containing the x and y acceleration from an accelerometer installed
on a cattle. There are missing days in this dataset.
The data set is sampled with 15 minutes samples. The data is as follows:
}
\keyword{datasets}
