% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df691b_1}
\alias{df691b_1}
\title{df691b_1 Activity Data Sets}
\format{
A data frame of 3 columns
\describe{
\item{datetime}{a POSIX formatted datetime}
\item{Motion.Index}{The motion index of the cow during the time sample}
\item{Steps}{The number of steps during the time sample}
}
}
\source{
Agroscope Tanikon
}
\usage{
df691b_1
}
\description{
A dataset containing the Motion index and steps count of a cow.
The data set is sampled with 15 minutes samples. The data is as follows:
}
\keyword{datasets}
