% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_act_info}
\alias{df_act_info}
\title{Outputs some information about the activity dataframe}
\usage{
df_act_info(df)
}
\arguments{
\item{df}{The dataframe containing the activity data}
}
\description{
Outputs some information about the activity dataframe
}
