% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfc.R
\name{dfc}
\alias{dfc}
\title{Computes the Degree of Function coupling for one variable in an activity dataset}
\usage{
dfc(
  data,
  activity = "Motion.Index",
  sampling = 15,
  save = FALSE,
  tag = NULL,
  outputdir = NULL,
  show_lsp_plot = TRUE
)
}
\arguments{
\item{data}{The activity data set}

\item{activity}{The name of the activity}

\item{sampling}{The sampling period of the data set in minutes}

\item{save}{TRUE to save results in text files, FALSE otherwise (if TRUE,
subsequent arguments (tag and outputdir) should be given)}

\item{tag}{The suffix of the saved files (see the parameter outputdir for
more details)}

\item{outputdir}{where the data are saved if save is TRUE. The names of saved
files will be(dfc_tag.txt and spec_tag.txt)}

\item{show_lsp_plot}{if TRUE, LSP plots will be shown}
}
\value{
A list containing 2 dataframe. DFC dataframe that contain the
results of a DFC computation and SPEC Dataframe that contains the result of
spectrum computation.
The DFC contains 3 columns:
** The date
** The DFC computed over 7 days (but we only extract the first 24 hours = 96 values)
** The Harmonic Part
Data are supposed to sampled with a specific smpling rate. It should be the same sampling rate
as in the given argument @sampling
Missing days are not permitted. If you have data with half day, it should be
removed.
Data are saved in the output directory odir
}
\description{
Computes the Degree of Function coupling for one variable in an activity dataset
}
\examples{
data("df516b_2", package = "digiRhythm")
df <- df516b_2
df <- remove_activity_outliers(df)
df_act_info(df)
activity = names(df)[2]
my_dfc <- dfc(df, activity , sampling = 15)
}
