% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dgm_periodicity}
\alias{dgm_periodicity}
\title{Returns the periodicity of a digiRhythm dataframe}
\usage{
dgm_periodicity(data)
}
\arguments{
\item{data}{a digiRhythm friendly dataframe}
}
\value{
returns a periodicity object of type xts.
}
\description{
Returns the periodicity of a digiRhythm dataframe
}
\examples{
data("df516b_2", package = "digiRhythm")
df <- df516b_2
dgm_periodicity(df)

}
