% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{levopt}
\alias{levopt}
\title{Returns the level given the p-value computed with pbaluev (2008). Copied from
the LOMB library.}
\usage{
levopt(Z, alpha, fmax, tm)
}
\arguments{
\item{Z}{the power of the frequency}

\item{alpha}{the significance level}

\item{fmax}{the maximum frequency in the spectrum}

\item{tm}{the time grid of the original time series}
}
\value{
Returns the level given the p-value computed with pbaluev (2008).
}
\description{
Returns the level given the p-value computed with pbaluev (2008). Copied from
the LOMB library.
}
