% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_activity_outliers}
\alias{remove_activity_outliers}
\title{Remove outliers from the data}
\usage{
remove_activity_outliers(df)
}
\arguments{
\item{df}{The dataframe containing the activity data}
}
\value{
return a dataframe where columns start the second one have undergone
an outlier removal.
}
\description{
Remove outliers from the data
}
