% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timedata}
\alias{timedata}
\title{timedata Dataset of start and end of day and night}
\format{
A data frame of 4 columns
\describe{
\item{day_start}{a POSIX formatted datetime}
\item{day_end}{a POSIX formatted datetime}
\item{night_start}{a POSIX formatted datetime}
\item{night_end}{a POSIX formatted datetime}
}
}
\source{
Johann Heinrich von Thünen- Institute of Organic Farming
}
\usage{
timedata
}
\description{
A dataset of start and endtime of the morningmilking and eveningmilking on a dairy farm.
}
\keyword{datasets}
