% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diurnality.R
\name{diurnality}
\alias{diurnality}
\title{Computes the diurnality index based on an activity dataframe}
\usage{
diurnality(data, activity, plot = FALSE)
}
\arguments{
\item{data}{a digiRhythm-friendly dataset}

\item{activity}{The number of non-useful lines to skip (lines to header)}

\item{plot}{if True, show the plot of diurnality index on the screen}
}
\value{
A dataframe with 2 col: date and diurnality index
}
\description{
Computes the diurnality index based on an activity dataframe
}
\examples{
data("df516b_2", package = "digiRhythm")
data <- df516b_2
data <- remove_activity_outliers(data)
activity = names(data)[2]
d_index <- diurnality(data, activity)
}
