% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pbaluev}
\alias{pbaluev}
\title{Returns p-value of a frequency peack according to pbaluev (2008) given Z,
fmax and tm. Copied from the LOMB library.}
\usage{
pbaluev(Z, fmax, tm)
}
\arguments{
\item{Z}{the power of the frequency}

\item{fmax}{the maximum frequency in the spectrum}

\item{tm}{the time grid of the original time series}
}
\value{
an intermediate calculation step needed to compute the p-value according to
pbaluev (2008).
}
\description{
Returns p-value of a frequency peack according to pbaluev (2008) given Z,
fmax and tm. Copied from the LOMB library.
}
