% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_v}
\alias{print_v}
\title{Print if Verbose is true}
\usage{
print_v(string, verbose)
}
\arguments{
\item{string}{The string to print}

\item{verbose}{if TRUE, print the string}
}
\value{
No return value. Prints the string concatenated with a verbose if the
latter is not NULL.
}
\description{
Print if Verbose is true
}
